+++ crates/qhyccd-rs/src/camera/parameters.rs
@@ -26,7 +26,7 @@ impl Camera {
     pub fn is_control_available(&self, control: Control) -> Option<u32> {
         match &self.backend {
             CameraBackend::Real { handle } => {
-                let handle = match read_lock!(handle, IsControlAvailableError { control }) {
+                let handle = match read_lock!(handle) {
                     Ok(handle) => handle,
                     Err(_) => return None,
                 };
@@ -72,7 +72,7 @@ impl Camera {
     pub fn get_parameter(&self, control: Control) -> Result<f64> {
         match &self.backend {
             CameraBackend::Real { handle } => {
-                let handle = read_lock!(handle, GetParameterError { control })?;
+                let handle = read_lock!(handle)?;
                 let res = unsafe { GetQHYCCDParam(handle, control as u32) };
                 if (res - QHYCCD_ERROR_F64).abs() < f64::EPSILON {
                     let error = GetParameterError { control };
@@ -133,7 +133,7 @@ impl Camera {
     pub fn get_parameter_min_max_step(&self, control: Control) -> Result<(f64, f64, f64)> {
         match &self.backend {
             CameraBackend::Real { handle } => {
-                let handle = read_lock!(handle, GetMinMaxStepError { control })?;
+                let handle = read_lock!(handle)?;
                 let mut min: f64 = 0.0;
                 let mut max: f64 = 0.0;
                 let mut step: f64 = 0.0;
@@ -186,7 +186,7 @@ impl Camera {
     pub fn set_parameter(&self, control: Control, value: f64) -> Result<()> {
         match &self.backend {
             CameraBackend::Real { handle } => {
-                let handle = read_lock!(handle, SetParameterError { error_code: 0 })?;
+                let handle = read_lock!(handle)?;
                 match unsafe { SetQHYCCDParam(handle, control as u32, value) } {
                     QHYCCD_SUCCESS => Ok(()),
                     error_code => {
@@ -258,7 +258,7 @@ impl Camera {
     pub fn is_cfw_plugged_in(&self) -> Result<bool> {
         match &self.backend {
             CameraBackend::Real { handle } => {
-                let handle = read_lock!(handle, IsCfwPluggedInError)?;
+                let handle = read_lock!(handle)?;
                 match unsafe { IsQHYCCDCFWPlugged(handle) } {
                     QHYCCD_SUCCESS => Ok(true),
                     QHYCCD_ERROR => Ok(false),